grpc_ext/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.2
requests
python-dotenv
pycryptodomex
httpx
